requirements.txt
flask
easyocr
Pillow
fpdf2
numpy
